+++ app/controllers/platform/tenants_controller.rb
@@ -84,7 +84,7 @@ module Platform
 
     def tenant_params
       params.require(:tenant).permit(
-        :name, :slug, :organization_id, :type, :status, :city, :country, :currency, :timezone
+        :name, :slug, :organization_id, :type, :status, :city, :address, :country, :currency, :timezone
       )
     end